fastapi/requirements.txt
fastapi
uvicorn
python-dotenv
motor
pymongo
itsdangerous>=2.0.1
starlette
pydantic[email]
email-validator
chromadb
PyJWT>=2.8.0
passlib[bcrypt]
aiomysql
langchain
langchain_community